requirement.txt
attrs==21.2.0
future==0.18.2
matplotlib==3.4.0
numpy==1.20.2
opencv-python==4.5.1.48
pandas==1.2.3
Pillow
tensorboard==2.0.2
tensorboard-plugin-wit==1.8.0
tensorboardX==2.1
torch==1.7.0+cu101
torchvision==0.8.1+cu101
tqdm==4.59.0
